Remove unit tests
At this stage the VM is undergoing too many changes for it to make sense
to have unit tests in place for features which may be changed radically
within the next hour. For this reason I'll be removing our current unit
tests and won't be writing new ones until the VM has become more mature.