Add `deinit` method to `BytecodeModule`
I'm not 100% happy with this implementation of `deinit` as it is
dependent on the caller passing the allocator which was used to allocate the
module's contents as an argument. Errors may occur if a different
allocator is passed as an argument instead. This could be solved by
storing the correct allocator as a struct member, but I'm unsure whether
this would be the cleanest way to resolve this issue.