Implement code blocks
Rather than storing a single array of instructions in our bytecode be
now store our instructions in multiple arrays known as blocks. This will
allow us to jump to blocks of code much more easily in the future and
will prove to be important when implementing subroutines and conditional
code execution.