Require bytecode module on VM initialization
Loading bytecode modules as a secondary step before executing
instructions seems like an unnecessary step given that loading a new
bytecode module without first deinitializing VM data is not a normal
use case for VMs. This commit removes this step by requiring a bytecode
module to be provided to the VM on initialization.