Add BlockIndex type in Value for block indices
Previously block indices were represented as regular integers. This
allowed programs to perform arithmetic operations on block indices as
well as pass arbitrary integers as block indices which could be very
dangerous. This commit adds a new type specifically for block indices.
With just this simple change the aforementioned scenarios will now be
caught as a type errors at runtime.