Implement debug print instruction
This commit adds a new `Print` instruction which pops a value off the
stack and writes it to `stderr` as a formatted string. This instruction
should eventually be replaced by equivalent instructions when I begin
working on a system interface for Assam and has only been added now for
debugging purposes.